split as module
